+++ src/bits.h
@@ -352,8 +352,8 @@ void bit_write_ENC (Bit_Chain *dat, Bit_Chain *hdl_dat, Bit_Chain *str_dat,
 int bit_search_sentinel (Bit_Chain *dat, const unsigned char sentinel[16]);
 void bit_write_sentinel (Bit_Chain *dat, const unsigned char sentinel[16]);
 
-void bit_chain_init (Bit_Chain *dat, const int size);
-void bit_chain_init_dat (Bit_Chain *restrict dat, const int size,
+void bit_chain_init (Bit_Chain *dat, const size_t size);
+void bit_chain_init_dat (Bit_Chain *restrict dat, const size_t size,
                          const Bit_Chain *restrict from_dat);
 void bit_chain_alloc (Bit_Chain *dat);
 void bit_chain_alloc_size (Bit_Chain *dat, const size_t size);
